Refactor SQL scripts and environment configuration for improved data type consistency and performance. Changed ID fields in multiple tables from DECIMAL(20,0) to BIGINT to enhance compatibility with JDBC lookups. Updated the run-sql.sh script to remove deprecated CDC server ID variables and streamline variable definitions for better clarity.

+++ flink-sync/sql/02_sync_id_mapping_incr.sql
@@ -19,8 +19,7 @@ SET 'table.exec.async-lookup.buffer-capacity' = '200';
 SET 'table.exec.async-lookup.timeout' = '60s';
 
 CREATE TABLE IF NOT EXISTS cdc_market_app (
-    -- market.application.id 多为 unsigned；DECIMAL 避免 LookupJoin ClassCast
-    id DECIMAL(20, 0),
+    id BIGINT,
     proc_time AS PROCTIME(),
     PRIMARY KEY (id) NOT ENFORCED
 ) WITH (
@@ -40,8 +39,8 @@ CREATE TABLE IF NOT EXISTS cdc_market_app (
 );
 
 CREATE TABLE IF NOT EXISTS cdc_user_data (
-    id DECIMAL(20, 0),
-    `userId` DECIMAL(20, 0),
+    id BIGINT,
+    `userId` BIGINT,
     proc_time AS PROCTIME(),
     PRIMARY KEY (id) NOT ENFORCED
 ) WITH (
@@ -61,7 +60,7 @@ CREATE TABLE IF NOT EXISTS cdc_user_data (
 );
 
 CREATE TABLE IF NOT EXISTS cdc_user (
-    id DECIMAL(20, 0),
+    id BIGINT,
     proc_time AS PROCTIME(),
     PRIMARY KEY (id) NOT ENFORCED
 ) WITH (
@@ -81,8 +80,7 @@ CREATE TABLE IF NOT EXISTS cdc_user (
 );
 
 CREATE TABLE IF NOT EXISTS dim_idmap_bundle (
-    -- 与视图裸 a.id 对齐；勿在视图 CAST，否则点查全表扫
-    app_row_id DECIMAL(20, 0),
+    app_row_id BIGINT,
     app_id BIGINT,
     mobile_norm STRING,
     mobile_token STRING,
@@ -108,8 +106,8 @@ CREATE TABLE IF NOT EXISTS dim_idmap_bundle (
 );
 
 CREATE TABLE IF NOT EXISTS dim_apps_by_user (
-    user_id DECIMAL(20, 0),
-    app_row_id DECIMAL(20, 0),
+    user_id BIGINT,
+    app_row_id BIGINT,
     PRIMARY KEY (user_id) NOT ENFORCED
 ) WITH (
     'connector' = 'jdbc',
